Bug 490475. Support the evaluation of C++14 constexpr functions

+++ qt/org.eclipse.cdt.qt.core/src/org/eclipse/cdt/internal/qt/core/index/QEnum.java
@@ -52,7 +52,7 @@ public class QEnum implements IQEnum {
 			this.name = enumerator.getName();
 
 			IValue val = enumerator.getValue();
-			this.ordinal = val == null ? null : val.numericalValue();
+			this.ordinal = val == null ? null : val.numericalValue().longValue();
 		}
 
 		@Override
